requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
